+++ src/devices/cooktop.test.ts
@@ -1,28 +1,26 @@
-// src/oven.test.ts
+// src/cooktop.test.ts
 
-const MATTER_PORT = 6024;
-const NAME = 'Oven';
+const MATTER_PORT = 6025;
+const NAME = 'Cooktop';
 const HOMEDIR = path.join('jest', NAME);
 
 import { rmSync } from 'node:fs';
 import path from 'node:path';
 
 import { jest } from '@jest/globals';
-import { AnsiLogger, LogLevel } from 'node-ansi-logger';
+import { AnsiLogger } from 'node-ansi-logger';
 // matter.js
 import { Endpoint, DeviceTypeId, VendorId, ServerNode, LogFormat as MatterLogFormat, LogLevel as MatterLogLevel, Environment, PositionTag } from '@matter/main';
 import { MdnsService } from '@matter/main/protocol';
 import { AggregatorEndpoint } from '@matter/main/endpoints/aggregator';
 import { RootEndpoint } from '@matter/main/endpoints/root';
-import { OvenCavityOperationalStateServer, OvenModeServer } from '@matter/main/behaviors';
-import { Identify, OnOff, OperationalState, OvenCavityOperationalState, OvenMode, PowerSource } from '@matter/main/clusters';
+import { Identify, OnOff, PowerSource, TemperatureControl, TemperatureMeasurement } from '@matter/main/clusters';
 
 // Matterbridge
 import { MatterbridgeEndpoint } from '../matterbridgeEndpoint.js';
-import { invokeBehaviorCommand } from '../matterbridgeEndpointHelpers.js';
 import { inspectError } from '../utils/error.js';
 
-import { MatterbridgeOvenCavityOperationalStateServer, MatterbridgeOvenModeServer, Oven } from './oven.js';
+import { Cooktop } from './cooktop.js';
 
 let loggerLogSpy: jest.SpiedFunction<typeof AnsiLogger.prototype.log>;
 let consoleLogSpy: jest.SpiedFunction<typeof console.log>;
@@ -55,9 +53,9 @@ describe('Matterbridge ' + NAME, () => {
   const environment = Environment.default;
   let server: ServerNode<ServerNode.RootEndpoint>;
   let aggregator: Endpoint<AggregatorEndpoint>;
-  let device: Oven;
-  let cabinet1: MatterbridgeEndpoint;
-  let cabinet2: MatterbridgeEndpoint;
+  let device: Cooktop;
+  let surface1: MatterbridgeEndpoint;
+  let surface2: MatterbridgeEndpoint;
 
   beforeAll(async () => {
     // Setup the matter environment
@@ -124,45 +122,39 @@ describe('Matterbridge ' + NAME, () => {
     expect(aggregator.lifecycle.isReady).toBeTruthy();
   });
 
-  test('create a oven device', async () => {
-    device = new Oven('Oven Test Device', 'OV123456');
+  test('create a cooktop device', async () => {
+    device = new Cooktop('Cooktop Test Device', 'CT123456');
     expect(device).toBeDefined();
-    expect(device.id).toBe('OvenTestDevice-OV123456');
+    expect(device.id).toBe('CooktopTestDevice-CT123456');
     expect(device.hasClusterServer(Identify.Cluster.id)).toBeTruthy();
     expect(device.hasClusterServer(PowerSource.Cluster.id)).toBeTruthy();
-    expect(device.hasClusterServer(OnOff.Cluster.id)).toBeFalsy();
-    expect(device.getAllClusterServerNames()).toEqual(['descriptor', 'matterbridge', 'identify', 'powerSource', 'fixedLabel']);
+    expect(device.hasClusterServer(OnOff.Cluster.id)).toBeTruthy();
+    expect(device.getAllClusterServerNames()).toEqual(['descriptor', 'matterbridge', 'identify', 'powerSource', 'onOff', 'fixedLabel']);
 
-    cabinet1 = device.addCabinet('Oven Test Cabinet Top', [{ mfgCode: null, namespaceId: PositionTag.Top.namespaceId, tag: PositionTag.Top.tag, label: PositionTag.Top.label }]);
-    expect(cabinet1).toBeDefined();
-    expect(cabinet1.id).toBe('OvenTestCabinetTop');
-    expect(cabinet1.hasClusterServer(OvenMode.Cluster.id)).toBeTruthy();
-    expect(cabinet1.hasClusterServer(OvenCavityOperationalState.Cluster.id)).toBeTruthy();
-    expect(cabinet1.getAllClusterServerNames()).toEqual(['descriptor', 'matterbridge', 'identify', 'temperatureControl', 'temperatureMeasurement', 'ovenMode', 'ovenCavityOperationalState']);
+    surface1 = device.addSurface('Surface Top Left', [
+      { mfgCode: null, namespaceId: PositionTag.Top.namespaceId, tag: PositionTag.Top.tag, label: PositionTag.Top.label },
+      { mfgCode: null, namespaceId: PositionTag.Left.namespaceId, tag: PositionTag.Left.tag, label: PositionTag.Left.label },
+    ]);
+    expect(surface1).toBeDefined();
+    expect(surface1.id).toBe('SurfaceTopLeft');
+    expect(surface1.hasClusterServer(OnOff.Cluster.id)).toBeTruthy();
+    expect(surface1.hasClusterServer(TemperatureControl.Cluster.id)).toBeTruthy();
+    expect(surface1.hasClusterServer(TemperatureMeasurement.Cluster.id)).toBeTruthy();
+    expect(surface1.getAllClusterServerNames()).toEqual(['descriptor', 'matterbridge', 'temperatureControl', 'temperatureMeasurement', 'onOff']);
 
-    cabinet2 = device.addCabinet(
-      'Oven Test Cabinet Bottom',
-      [{ mfgCode: null, namespaceId: PositionTag.Bottom.namespaceId, tag: PositionTag.Bottom.tag, label: PositionTag.Bottom.label }],
-      3,
-      [
-        { label: 'Convection', mode: 1, modeTags: [{ value: OvenMode.ModeTag.Convection }] },
-        { label: 'Clean', mode: 2, modeTags: [{ value: OvenMode.ModeTag.Clean }] },
-        { label: 'Steam', mode: 3, modeTags: [{ value: OvenMode.ModeTag.Steam }] },
-      ],
-      2,
-      ['180°', '190°', '200°'],
-      OperationalState.OperationalStateEnum.Running,
-      1,
-      ['pre-heating', 'pre-heated', 'cooling down'],
-    );
-    expect(cabinet2).toBeDefined();
-    expect(cabinet2.id).toBe('OvenTestCabinetBottom');
-    expect(cabinet2.hasClusterServer(OvenMode.Cluster.id)).toBeTruthy();
-    expect(cabinet2.hasClusterServer(OvenCavityOperationalState.Cluster.id)).toBeTruthy();
-    expect(cabinet2.getAllClusterServerNames()).toEqual(['descriptor', 'matterbridge', 'identify', 'temperatureControl', 'temperatureMeasurement', 'ovenMode', 'ovenCavityOperationalState']);
+    surface2 = device.addSurface('Surface Top Right', [
+      { mfgCode: null, namespaceId: PositionTag.Top.namespaceId, tag: PositionTag.Top.tag, label: PositionTag.Top.label },
+      { mfgCode: null, namespaceId: PositionTag.Right.namespaceId, tag: PositionTag.Right.tag, label: PositionTag.Right.label },
+    ]);
+    expect(surface2).toBeDefined();
+    expect(surface2.id).toBe('SurfaceTopRight');
+    expect(surface2.hasClusterServer(OnOff.Cluster.id)).toBeTruthy();
+    expect(surface2.hasClusterServer(TemperatureControl.Cluster.id)).toBeTruthy();
+    expect(surface2.hasClusterServer(TemperatureMeasurement.Cluster.id)).toBeTruthy();
+    expect(surface2.getAllClusterServerNames()).toEqual(['descriptor', 'matterbridge', 'temperatureControl', 'temperatureMeasurement', 'onOff']);
   });
 
-  test('add a oven device', async () => {
+  test('add a cooktop device', async () => {
     expect(server).toBeDefined();
     expect(device).toBeDefined();
     try {
@@ -171,12 +163,9 @@ describe('Matterbridge ' + NAME, () => {
       inspectError(device.log, `Error adding device ${device.deviceName}`, error);
       return;
     }
-    expect(server.parts.has('OvenTestDevice-OV123456')).toBeTruthy();
+    expect(server.parts.has('CooktopTestDevice-CT123456')).toBeTruthy();
     expect(server.parts.has(device)).toBeTruthy();
     expect(device.lifecycle.isReady).toBeTruthy();
-
-    expect(loggerLogSpy).toHaveBeenCalledWith(LogLevel.INFO, `MatterbridgeOvenModeServer initialized`);
-    expect(loggerLogSpy).toHaveBeenCalledWith(LogLevel.INFO, `MatterbridgeOvenCavityOperationalStateServer initialized: setting operational state to Stopped and operational error to No error`);
   });
 
   test('start the server node', async () => {
@@ -215,12 +204,12 @@ describe('Matterbridge ' + NAME, () => {
       expect(attributeId).toBeGreaterThanOrEqual(0);
       attributes.push({ clusterName, clusterId, attributeName, attributeId, attributeValue });
     });
-    expect(attributes.length).toBe(39); // 39 attributes for the oven device
+    expect(attributes.length).toBe(45); // 45 attributes for the oven device
   });
 
-  test('cabinet1 forEachAttribute', async () => {
+  test('surface1 forEachAttribute', async () => {
     const attributes: { clusterName: string; clusterId: number; attributeName: string; attributeId: number; attributeValue: any }[] = [];
-    cabinet1.forEachAttribute((clusterName, clusterId, attributeName, attributeId, attributeValue) => {
+    surface1.forEachAttribute((clusterName, clusterId, attributeName, attributeId, attributeValue) => {
       expect(clusterName).toBeDefined();
       expect(typeof clusterName).toBe('string');
 
@@ -236,12 +225,12 @@ describe('Matterbridge ' + NAME, () => {
       expect(attributeId).toBeGreaterThanOrEqual(0);
       attributes.push({ clusterName, clusterId, attributeName, attributeId, attributeValue });
     });
-    expect(attributes.length).toBe(51); // 51 attributes for the cabinet1 device
+    expect(attributes.length).toBe(32); // 32 attributes for the surface1 device
   });
 
-  test('cabinet2 forEachAttribute', async () => {
+  test('surface2 forEachAttribute', async () => {
     const attributes: { clusterName: string; clusterId: number; attributeName: string; attributeId: number; attributeValue: any }[] = [];
-    cabinet2.forEachAttribute((clusterName, clusterId, attributeName, attributeId, attributeValue) => {
+    surface2.forEachAttribute((clusterName, clusterId, attributeName, attributeId, attributeValue) => {
       expect(clusterName).toBeDefined();
       expect(typeof clusterName).toBe('string');
 
@@ -257,46 +246,7 @@ describe('Matterbridge ' + NAME, () => {
       expect(attributeId).toBeGreaterThanOrEqual(0);
       attributes.push({ clusterName, clusterId, attributeName, attributeId, attributeValue });
     });
-    expect(attributes.length).toBe(51); // 51 attributes for the cabinet2 device
-  });
-
-  test('invoke MatterbridgeOvenModeServer commands', async () => {
-    expect(cabinet1.behaviors.has(OvenModeServer)).toBeTruthy();
-    expect(cabinet1.behaviors.has(MatterbridgeOvenModeServer)).toBeTruthy();
-    expect(cabinet1.behaviors.elementsOf(OvenModeServer).commands.has('changeToMode')).toBeTruthy();
-    expect((cabinet1 as any).state['ovenMode'].acceptedCommandList).toEqual([0]);
-    expect((cabinet1 as any).state['ovenMode'].generatedCommandList).toEqual([1]);
-
-    // Change to mode 2
-    jest.clearAllMocks();
-    await invokeBehaviorCommand(cabinet1, 'ovenMode', 'changeToMode', { newMode: 2 });
-    expect(loggerLogSpy).toHaveBeenCalledWith(LogLevel.INFO, `MatterbridgeOvenModeServer: changeToMode (endpoint OvenTestCabinetTop.3) called with mode 2 = Convection`);
-
-    // Change to mode 15
-    jest.clearAllMocks();
-    await invokeBehaviorCommand(cabinet1, 'ovenMode', 'changeToMode', { newMode: 15 });
-    expect(loggerLogSpy).toHaveBeenCalledWith(LogLevel.ERROR, `MatterbridgeOvenModeServer: changeToMode (endpoint OvenTestCabinetTop.3) called with invalid mode 15`);
-  });
-
-  test('invoke MatterbridgeOvenCavityOperationalStateServer commands', async () => {
-    expect(cabinet1.behaviors.has(OvenCavityOperationalStateServer)).toBeTruthy();
-    expect(cabinet1.behaviors.has(MatterbridgeOvenCavityOperationalStateServer)).toBeTruthy();
-    expect(cabinet1.behaviors.elementsOf(OvenCavityOperationalStateServer).commands.has('stop')).toBeTruthy();
-    expect(cabinet1.behaviors.elementsOf(OvenCavityOperationalStateServer).commands.has('start')).toBeTruthy();
-    expect(cabinet1.behaviors.elementsOf(OvenCavityOperationalStateServer).commands.has('pause')).toBeFalsy();
-    expect(cabinet1.behaviors.elementsOf(OvenCavityOperationalStateServer).commands.has('resume')).toBeFalsy();
-    expect((cabinet1 as any).state['ovenCavityOperationalState'].acceptedCommandList).toEqual([1, 2]);
-    expect((cabinet1 as any).state['ovenCavityOperationalState'].generatedCommandList).toEqual([4]);
-
-    // Change to mode 2
-    jest.clearAllMocks();
-    await invokeBehaviorCommand(cabinet1, 'ovenCavityOperationalState', 'start', { newMode: 2 });
-    expect(loggerLogSpy).toHaveBeenCalledWith(LogLevel.INFO, `MatterbridgeOvenCavityOperationalStateServer: start (endpoint OvenTestCabinetTop.3) called setting operational state to Running and operational error to No error`);
-
-    // Change to mode 15
-    jest.clearAllMocks();
-    await invokeBehaviorCommand(cabinet1, 'ovenCavityOperationalState', 'stop', { newMode: 15 });
-    expect(loggerLogSpy).toHaveBeenCalledWith(LogLevel.INFO, `MatterbridgeOvenCavityOperationalStateServer: stop (endpoint OvenTestCabinetTop.3) called setting operational state to Stopped and operational error to No error`);
+    expect(attributes.length).toBe(32); // 32 attributes for the surface2 device
   });
 
   test('close the server node', async () => {