+++ src/devices/export.ts
@@ -5,6 +5,7 @@ export * from './extractorHood.js';
 export * from './dishwasher.js';
 export * from './microwaveOven.js';
 export * from './oven.js';
+export * from './cooktop.js';
 
 export * from './waterHeater.js';
 export * from './evse.js';